Cleaning and using Boxing instead of Boxing2
